feature/1640: fix mobile header
