+++ src/assets/sass/_header.sass
@@ -216,19 +216,22 @@
 
 // header-media
 .header-media
-	display: none
+	width: 100%
+	height: $mobileHeaderHeight
 
-	top: 0
-	position: sticky
-
-	z-index: 10
+	display: flex
+	align-items: center
 
 	background-color: $backgroundColor
 
-	padding: 20px 0
-	margin-bottom: 30px
+	position: sticky
+	top: 0
+	z-index: 10
 
-	box-shadow: 0px 2px 20px 0px rgba(0, 0, 0, 0.15)
+	transition: background-color .2s ease
+
+	&--transparent
+		background-color: rgba($backgroundColor, 0)
 
 	&-wrapper
 		width: 100%
@@ -241,11 +244,14 @@
 	&-icon-group
 		display: flex
 		align-items: center
+		gap: 20px
 
 	&-icon
 		position: relative
-
-		margin: 0 10px
+		width: 28px
+		height: 28px
+		+flexCenter
+		flex-shrink: 0
 
 		&__count
 			cursor: pointer
@@ -273,23 +279,20 @@
 
 	&-logo
 		&__image
-			width: 120px
+			height: 20px
 
 	&-modal-menu
 		width: 100%
 
-		top: 65px
-		position: absolute
+		position: fixed
+		top: $mobileHeaderHeight
+		height: calc(100% - #{$mobileHeaderHeight})
 
 		background-color: $backgroundColor
 
 		opacity: 0
 		pointer-events: none
 		visibility: hidden
-
-		padding-bottom: 75px
-
-		height: 100vh
 
 		overflow-y: auto
 
@@ -411,14 +414,3 @@
 
 				&:first-child
 					margin-bottom: 10px
-
-@media screen and ( max-width: 1250px )
-	.header
-		display: none
-
-		&-hover-menu-wrapper-wrapper
-			display: none !important
-
-	.header-media
-		display: block
-		margin-bottom: 0
