+++ src/assets/sass/_variables.sass
@@ -28,3 +28,6 @@ $grey6: #808080
 $grey8: #303030
 $grey9: #2C2C2C
 $grey10: #282828
+
+// sizes
+$mobileHeaderHeight: 50px
